Moved tree implementations into their own module
