+++ src/parser/mod.rs
@@ -4,7 +4,7 @@ A parser combinator, inspired by nom.
 This parser combinator passes a context into the function, which includes the string being parsed. This supports resolving context-based constructs such as general entities and XML Namespaces.
 */
 
-use crate::intmuttree::{ExtDTDresolver, DTD};
+use crate::trees::intmuttree::{ExtDTDresolver, DTD};
 use crate::xdmerror::{Error, ErrorKind};
 use std::collections::HashMap;
 use std::fmt;
