+++ src/trees/mod.rs
@@ -1,0 +1,3 @@
+//! Various implementations of tree data structures.
+
+pub mod intmuttree;
